renamed some card components
