+++ src/components/BracketCard/CreateBracketCard.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import BracketCard from "./BracketCard";
+import Card from "./Card";
 import createBracketPic from "../../assets/images/createBracket.png";
 import { CreateBracketModal } from "./CreateBracketModal";
 
@@ -7,7 +7,7 @@ const CreateBracketCard = ({ userId }) => {
     const [showModal, setShowModal] = useState(false);
     return (
         <div>
-            <BracketCard
+            <Card
                 image={createBracketPic}
                 imageAlt="Plus sign"
                 cardText={"Create Bracket"}
